# swagger.yaml
openapi: 3.0.0
paths:
...
  /auth/login:
    post:
      operationId: AuthController_signIn
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignInDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
      tags: &ref_10
        - auth
    get:
      operationId: AuthController_root
      parameters: []
      responses:
        '200':
          description: ''
      tags: *ref_10
  /auth/register:
    post:
      operationId: AuthController_signUp
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags: *ref_10
  /auth/create_admin:
    post:
      operationId: AuthController_createAdmin
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags: *ref_10
      security:
        - Authorization: []
  /auth/users:
    get:
      operationId: AuthController_getAllUsers
      parameters: []
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
      tags: *ref_10
      security:
        - Authorization: []
  /auth/users/{id}:
    get:
      operationId: AuthController_getUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
      tags: *ref_10
      security:
        - Authorization: []
    put:
      operationId: AuthController_updateUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserDto'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      tags: *ref_10
      security:
        - Authorization: []
    delete:
      operationId: AuthController_deleteUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
      tags: *ref_10
      security:
        - Authorization: []
info:
  title: UG Examine
  description: UG Examine API description
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    User:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        fullName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        role:
          enum:
            - student
            - staff
            - admin
          type: string
      required:
        - username
        - password
        - email
        - phoneNumber
        - role
...
    SignInDto:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password
    CreateUserDto:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        fullName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string
        role:
          enum:
            - student
            - staff
            - admin
          type: string
      required:
        - username
        - password
        - email
        - phoneNumber
        - role
    UpdateUserDto:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
        phoneNumber:
          type: string